+++ example/lib/main.dart
@@ -33,8 +33,6 @@ class MyHomePage extends StatefulWidget {
 
 class _MyHomePageState extends State<MyHomePage> {
   final padding = 8.0;
-  final spacing = 0.0;
-  final itemWidth = 380.0;
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -44,23 +42,43 @@ class _MyHomePageState extends State<MyHomePage> {
       ),
       body: SingleChildScrollView(
         child: FlexibleWrap(
-          length: 4, // Number of children to display
           isOneRowExpanded: true,
-          spacing: spacing,
-          builder: (int index) {
+          spacing: 12.0,
+          children: List.generate(20, (int index) {
             return Padding(
               padding: EdgeInsets.all(padding),
               child: Container(
                 height: 100,
+                width: 300,
                 decoration: BoxDecoration(
                     color: Colors.blue,
                     borderRadius: BorderRadius.circular(8.0)),
-                child: Center(child: Text('Item $index')),
+                child: const Center(
+                  child: ListTile(
+                    title: Text(
+                      "Lorem Ipsum is simply dummy text",
+                      style: TextStyle(color: Colors.white),
+                      overflow: TextOverflow.ellipsis,
+                    ),
+                    subtitle: Text(
+                      "Lorem Ipsum has been the industry's standard",
+                      style: TextStyle(color: Colors.white),
+                      overflow: TextOverflow.ellipsis,
+                    ),
+                    leading: Icon(
+                      Icons.insert_emoticon,
+                      color: Colors.white,
+                      size: 60.0,
+                    ),
+                    trailing: Icon(
+                      Icons.favorite,
+                      color: Colors.white,
+                    ),
+                  ),
+                ),
               ),
             );
-          },
-          itemWidth: itemWidth, // Width of each item
-          alignment: WrapAlignment.start, // Alignment of children within a run
+          }).toList(),
         ),
       ),
     );